Initial design of the scheduler api design
