+++ src/definition/accessors/IModifyCreator.ts
@@ -1,0 +1,32 @@
+import { IMessage } from '../messages';
+import { IRoom } from '../rooms';
+import { IMessageBuilder } from './IMessageBuilder';
+import { IRoomBuilder } from './IRoomBuilder';
+
+export interface IModifyCreator {
+    /**
+     * Starts the process for building a new message object.
+     *
+     * @param data (optional) the initial data to pass into the builder,
+     *          the `id` property will be ignored
+     * @return an IMessageBuilder instance
+     */
+    startMessage(data?: IMessage): IMessageBuilder;
+
+    /**
+     * Starts the process for building a new room.
+     *
+     * @param data (optional) the initial data to pass into the builder,
+     *          the `id` property will be ignored
+     * @return an IRoomBuilder instance
+     */
+    startRoom(data?: IRoom): IRoomBuilder;
+
+    /**
+     * Finishes the creating process, saving the object to the database.
+     *
+     * @param builder the builder instance
+     * @return the resulting `id` of the resulting object
+     */
+    finish(builder: IMessageBuilder | IRoomBuilder): Promise<string>;
+}
