+++ src/definition/accessors/ISchedulerExtend.ts
@@ -1,0 +1,8 @@
+import { IJob } from '../scheduler';
+import { IJobBuilder } from './IJobBuilder';
+
+export interface ISchedulerExtend {
+    scheduleJob(job: IJob | IJobBuilder): Promise<void>;
+
+    getJobBuilder(job?: IJob): IJobBuilder;
+}
