+++ src/server/accessors/ConfigurationExtend.ts
@@ -2,6 +2,7 @@ import {
     IApiExtend,
     IConfigurationExtend,
     IHttpExtend,
+    ISchedulerExtend,
     ISettingsExtend,
     ISlashCommandsExtend,
 } from '../../definition/accessors';
@@ -11,11 +12,13 @@ export class ConfigurationExtend implements IConfigurationExtend {
     public readonly settings: ISettingsExtend;
     public readonly slashCommands: ISlashCommandsExtend;
     public readonly api: IApiExtend;
+    public readonly scheduler: ISchedulerExtend;
 
-    constructor(https: IHttpExtend, sets: ISettingsExtend, cmds: ISlashCommandsExtend, api: IApiExtend) {
+    constructor(https: IHttpExtend, sets: ISettingsExtend, cmds: ISlashCommandsExtend, api: IApiExtend, scheduler: ISchedulerExtend) {
         this.http = https;
         this.settings = sets;
         this.slashCommands = cmds;
         this.api = api;
+        this.scheduler = scheduler;
     }
 }
