+++ src/server/accessors/index.ts
@@ -18,6 +18,7 @@ import { Reader } from './Reader';
 import { RoomBuilder } from './RoomBuilder';
 import { RoomExtender } from './RoomExtender';
 import { RoomRead } from './RoomRead';
+import { SchedulerExtend } from './SchedulerExtend';
 import { ServerSettingRead } from './ServerSettingRead';
 import { ServerSettingsModify } from './ServerSettingsModify';
 import { SettingRead } from './SettingRead';
@@ -47,6 +48,7 @@ export {
     RoomBuilder,
     RoomExtender,
     RoomRead,
+    SchedulerExtend,
     ServerSettingRead,
     ServerSettingsModify,
     SettingRead,
