+++ tests/server/accessors/ConfigurationExtend.spec.ts
@@ -1,5 +1,5 @@
 import { Expect, SetupFixture, Test } from 'alsatian';
-import { IApiExtend, IHttpExtend, ISettingsExtend, ISlashCommandsExtend } from '../../../src/definition/accessors';
+import { IApiExtend, IHttpExtend, ISchedulerExtend, ISettingsExtend, ISlashCommandsExtend } from '../../../src/definition/accessors';
 
 import { ConfigurationExtend } from '../../../src/server/accessors';
 
@@ -8,6 +8,7 @@ export class ConfigurationExtendTestFixture {
     private se: ISettingsExtend;
     private sce: ISlashCommandsExtend;
     private api: IApiExtend;
+    private sch: ISchedulerExtend;
 
     @SetupFixture
     public setupFixture() {
@@ -15,13 +16,14 @@ export class ConfigurationExtendTestFixture {
         this.se = {} as ISettingsExtend;
         this.sce = {} as ISlashCommandsExtend;
         this.api = {} as IApiExtend;
+        this.sch = {} as ISchedulerExtend;
     }
 
     @Test()
     public useConfigurationExtend() {
-        Expect(() => new ConfigurationExtend(this.he, this.se, this.sce, this.api)).not.toThrow();
+        Expect(() => new ConfigurationExtend(this.he, this.se, this.sce, this.api, this.sch)).not.toThrow();
 
-        const se = new ConfigurationExtend(this.he, this.se, this.sce, this.api);
+        const se = new ConfigurationExtend(this.he, this.se, this.sce, this.api, this.sch);
         Expect(se.http).toBeDefined();
         Expect(se.settings).toBeDefined();
         Expect(se.slashCommands).toBeDefined();
